% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{gpo_package_granules_summary}
\alias{gpo_package_granules_summary}
\title{Given a package id and a granule id, return a metadata summary for the granule.}
\usage{
gpo_package_granules_summary(package_id, granule_id)
}
\arguments{
\item{package_id}{String. The Package Id. Ex: CREC-2018-01-04}

\item{granule_id}{String. The granule ID, e.g. CREC-2018-01-04-pt1-PgD7-2}
}
\value{
A single row tibble
}
\description{
Given a package id and a granule id, return a metadata summary for the granule.
}
\examples{
\dontshow{if (govinfoR::has_govinfo_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

gpo_package_granules_summary("CREC-2018-01-04", "CREC-2018-01-04-pt1-PgD7-2")
\dontshow{\}) # examplesIf}
}
