% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{gpo_package_summary}
\alias{gpo_package_summary}
\title{Given a GPO package id, return summary metadata for that package.}
\usage{
gpo_package_summary(package_id)
}
\arguments{
\item{package_id}{String. The Package Id. Ex: CREC-2018-01-04}
}
\value{
A single row tibble.
}
\description{
Given a GPO package id, return summary metadata for that package.
}
\examples{
set_govinfo_key("DEMO_KEY")
gpo_package_summary("CREC-2018-01-04")
}
