% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/related.R
\name{gpo_related}
\alias{gpo_related}
\title{gpo_related()}
\usage{
gpo_related(
  access_id,
  collection = NULL,
  granule_class = NULL,
  sub_granule_class = NULL
)
}
\arguments{
\item{access_id}{GPO package id or granule id}

\item{collection}{Collection to search for relationships. Must be one of BILLS, BILLSTATUS,
PLAW, FR, or CHRG.}

\item{granule_class}{A type of granule.}

\item{sub_granule_class}{A type of subgranule.}
}
\value{
A tibble.
}
\description{
gpo_related()
}
\examples{
\dontshow{if (govinfoR::has_govinfo_key()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

gpo_related("BILLS-116hr748enr")
gpo_related("BILLS-116hr748enr", collection = "BILLS")
\dontshow{\}) # examplesIf}
}
