% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govinfo_key.R
\name{key}
\alias{key}
\alias{has_govinfo_key}
\alias{get_govinfo_key}
\title{Check or Get GovInfo API Key}
\usage{
has_govinfo_key()

get_govinfo_key()
}
\value{
logical if \code{has}, key if \code{get}
}
\description{
Adapted, with gratitude, from Christopher Kenny's
\href{https://github.com/christopherkenny/congress/blob/main/R/congress_key.R}{congress}
package
}
\examples{
has_govinfo_key()
}
