% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{next_req}
\alias{next_req}
\title{next_req}
\usage{
next_req(resp, req)
}
\arguments{
\item{resp}{Callback response.}

\item{req}{Callback response.}
}
\description{
A callback for use with httr2::req_perform_iterative
}
