% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/govinfo_key.R
\name{set_govinfo_key}
\alias{set_govinfo_key}
\title{Adds api.data.gov key to .Renviron}
\usage{
set_govinfo_key(key, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{Character. API key to add.}

\item{overwrite}{Defaults to FALSE. Boolean. Should existing \code{GOVINFO_KEY} in .Renviron  be overwritten?}

\item{install}{Defaults to FALSE. Boolean. Should \code{key} be added to '~/.Renviron' file?}
}
\value{
key, invisibly
}
\description{
Adds api.data.gov key to .Renviron
}
\examples{
\dontrun{
  set_govinfo_key("DEMO_KEY")
}
}
