% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{today}
\alias{today}
\title{today}
\usage{
today()
}
\value{
String. Represents the earliest possible UTC time for today's date.
}
\description{
A helper for specifying time stamps to the GovInfo API
}
\examples{
today()

}
