% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tomorrow}
\alias{tomorrow}
\title{tomorrow}
\usage{
tomorrow()
}
\value{
String. Represents the earliest possible UTC time for tomorrow's date.
}
\description{
A helper for specifying time stamps to the GovInfo API
}
\examples{
tomorrow()

}
