% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yesterday}
\alias{yesterday}
\title{yesterday}
\usage{
yesterday()
}
\value{
String. Represents the earliest possible UTC time for the previous day.
}
\description{
A helper for specifying time stamps to the GovInfo API
}
\examples{
yesterday()

}
